Version 4.1.0 (#169)
* Adding #118 #126 advanced panel with FFmpeg filters (thanks to Marco Ravich and remlap)
* Adding #114 info panel that shows source details (thanks to leonardyan)
* Adding #117 support for text based burn in subtitles
* Adding #152 option to switch back to old bubbly buttons (thanks to GitKike98)
* Adding logging level selector in settings for GUI
* Adding button to extract text based subtitles
* Adding dual pass option for ffmpeg options (thanks to Catatau)
* Adding #163 elapsed time in encoding panel (thanks to Benedicte Emilie Brækken)
* Adding super secret DEVMODE environment variable
* Adding #168 Allow decimal CRF values as Custom values (thanks to Etz)
* Changing all builds to use Python 3.8, removing idea of "legacy" windows until later time
* Fixing #156 Copy was broken due to ffmpeg extras not being passed (thanks to leonardyan)
* Fixing #158 Translation and icon of tab name "Quality" lost when output codec switched (thanks to leonardyan)
* Fixing #120 Custom bitrate forgets to add a "k" (thanks to seighail3)
* Fixing #162 Main window was not resizeable (thanks to bmcassagne)
* Fixing #165 Custom Bitrate loaded from Profile (thanks to Etz)
* Fixing #154 Overloading the term "Profile" for languages (thanks to leonardyan)
* Fixing snap packaged, and other symlinked FFmpeg installs were not linked correctly (thanks to Catatau)
* Fixing loading video could encounter a bug and crash fastflix (thanks to Jan)
* Fixing being able to set landscape cover
* Fixing resetting covers from queue

+++ fastflix/widgets/main.py
@@ -8,7 +8,6 @@ import os
 import secrets
 import shutil
 import time
-from dataclasses import asdict, dataclass, field
 from datetime import timedelta
 from pathlib import Path
 from typing import Tuple, Union
@@ -16,6 +15,7 @@ from typing import Tuple, Union
 import pkg_resources
 import reusables
 from box import Box
+from pydantic import BaseModel, Field
 from qtpy import QtCore, QtGui, QtWidgets
 
 from fastflix.encoders.common import helpers
@@ -41,8 +41,8 @@ from fastflix.resources import (
     video_playlist_icon,
 )
 from fastflix.shared import error_message, time_to_number, yes_no_message
+from fastflix.widgets.background_tasks import SubtitleFix, ThumbnailCreator
 from fastflix.widgets.progress_bar import ProgressBar, Task
-from fastflix.widgets.background_tasks import ThumbnailCreator, SubtitleFix
 from fastflix.widgets.video_options import VideoOptions
 
 logger = logging.getLogger("fastflix")
@@ -52,30 +52,33 @@ root = os.path.abspath(os.path.dirname(__file__))
 only_int = QtGui.QIntValidator()
 
 
-@dataclass
-class CropWidgets:
+class CropWidgets(BaseModel):
     top: QtWidgets.QLineEdit = None
     bottom: QtWidgets.QLineEdit = None
     left: QtWidgets.QLineEdit = None
     right: QtWidgets.QLineEdit = None
 
+    class Config:
+        arbitrary_types_allowed = True
 
-@dataclass
-class ScaleWidgets:
+
+class ScaleWidgets(BaseModel):
     width: QtWidgets.QLineEdit = None
     height: QtWidgets.QLineEdit = None
     keep_aspect: QtWidgets.QCheckBox = None
 
+    class Config:
+        arbitrary_types_allowed = True
 
-@dataclass
-class MainWidgets:
+
+class MainWidgets(BaseModel):
     start_time: QtWidgets.QLineEdit = None
     end_time: QtWidgets.QLineEdit = None
     video_track: QtWidgets.QComboBox = None
     rotate: QtWidgets.QComboBox = None
     flip: QtWidgets.QComboBox = None
-    crop: CropWidgets = field(default_factory=CropWidgets)
-    scale: ScaleWidgets = field(default_factory=ScaleWidgets)
+    crop: CropWidgets = Field(default_factory=CropWidgets)
+    scale: ScaleWidgets = Field(default_factory=ScaleWidgets)
     remove_metadata: QtWidgets.QCheckBox = None
     chapters: QtWidgets.QCheckBox = None
     fast_time: QtWidgets.QComboBox = None
@@ -86,6 +89,9 @@ class MainWidgets:
     remove_hdr: QtWidgets.QCheckBox = None
     video_title: QtWidgets.QLineEdit = None
     profile_box: QtWidgets.QComboBox = None
+
+    class Config:
+        arbitrary_types_allowed = True
 
     def items(self):
         for key in dir(self):
@@ -112,7 +118,7 @@ class Main(QtWidgets.QWidget):
         super().__init__(parent)
         self.app: FastFlixApp = app
         self.container = parent
-        self.video: Video = Video(Path(), 0, 0, 0)
+        self.video: Video = Video(source=Path(), width=0, height=0, duration=0)
 
         self.initialized = False
         self.loading_video = True
@@ -226,7 +232,7 @@ class Main(QtWidgets.QWidget):
         # add_profile.setFixedSize(QtCore.QSize(40, 40))
         add_profile.setFixedHeight(40)
         add_profile.setIconSize(QtCore.QSize(22, 22))
-        add_profile.setToolTip(t("Profile"))
+        add_profile.setToolTip(t("Profile_newprofiletooltip"))
         add_profile.setLayoutDirection(QtCore.Qt.RightToLeft)
         add_profile.clicked.connect(lambda: self.container.new_profile())
 
@@ -335,7 +341,7 @@ class Main(QtWidgets.QWidget):
 
         self.widgets.remove_hdr = QtWidgets.QCheckBox(t("Remove HDR"))
         self.widgets.remove_hdr.setChecked(False)
-        self.widgets.remove_hdr.toggled.connect(self.encoder_settings_update)
+        self.widgets.remove_hdr.toggled.connect(self.hdr_update)
         self.widgets.remove_hdr.setToolTip(
             f"{t('Convert BT2020 colorspace into bt709')}\n"
             f"{t('WARNING: This will take much longer and result in a larger file')}"
@@ -450,12 +456,12 @@ class Main(QtWidgets.QWidget):
         return self.widgets.rotate
 
     def rotation_to_transpose(self):
-        mapping = {0: None, 1: 1, 2: 4, 3: 2}
+        mapping = {0: 0, 1: 1, 2: 4, 3: 2}
         return mapping[self.widgets.rotate.currentIndex()]
 
     def transpose_to_rotation(self, transpose):
-        mapping = {None: 0, 1: 1, 4: 2, 2: 3}
-        return mapping[int(transpose) if transpose else None]
+        mapping = {0: 0, 1: 1, 4: 2, 2: 3}
+        return mapping[int(transpose)]
 
     def change_output_types(self):
         self.widgets.convert_to.clear()
@@ -464,7 +470,9 @@ class Main(QtWidgets.QWidget):
             if getattr(plugin, "icon", False):
                 self.widgets.convert_to.setItemIcon(i, QtGui.QIcon(plugin.icon))
         self.widgets.convert_to.setFont(QtGui.QFont("helvetica", 10, weight=57))
-        self.widgets.convert_to.setIconSize(QtCore.QSize(40, 40))
+        self.widgets.convert_to.setIconSize(
+            QtCore.QSize(40, 40) if self.app.fastflix.config.flat_ui else QtCore.QSize(35, 35)
+        )
 
     def init_encoder_drop_down(self):
         layout = QtWidgets.QHBoxLayout()
@@ -739,7 +747,14 @@ class Main(QtWidgets.QWidget):
         self.output_video_path_widget.setText(self.generate_output_filename)
         self.output_video_path_widget.setDisabled(False)
         self.output_path_button.setDisabled(False)
-        self.update_video_info()
+        try:
+            self.update_video_info()
+        except Exception:
+            logger.exception(f"Could not load video {self.input_video}")
+            self.video_path_widget.setText("")
+            self.output_video_path_widget.setText("")
+            self.output_video_path_widget.setDisabled(True)
+            self.output_path_button.setDisabled(True)
         self.page_update()
 
     @property
@@ -767,12 +782,17 @@ class Main(QtWidgets.QWidget):
         start_pos = self.start_time or self.app.fastflix.current_video.duration // 10
 
         blocks = math.ceil((self.app.fastflix.current_video.duration - start_pos) / 5)
+        if blocks < 1:
+            blocks = 1
 
         times = [
             x
             for x in range(int(start_pos), int(self.app.fastflix.current_video.duration), blocks)
             if x < self.app.fastflix.current_video.duration
         ][:4]
+
+        if not times:
+            return
 
         self.app.processEvents()
         result_list = []
@@ -931,7 +951,6 @@ class Main(QtWidgets.QWidget):
         except (ValueError, AssertionError):
             self.scale_updating = False
             return logger.warning(t("Invalid width"))
-            # return self.scale_warning_message.setText("Invalid main_width")
 
         if scale_width % 2:
             self.scale_updating = False
@@ -1217,7 +1236,7 @@ class Main(QtWidgets.QWidget):
         if not self.input_video or self.loading_video:
             return
 
-        settings = asdict(self.app.fastflix.current_video.video_settings)
+        settings = self.app.fastflix.current_video.video_settings.dict()
 
         if (
             self.app.fastflix.current_video.video_settings.video_encoder_settings.pix_fmt == "yuv420p10le"
@@ -1352,6 +1371,10 @@ class Main(QtWidgets.QWidget):
 
     def encoder_settings_update(self):
         self.video_options.settings_update()
+
+    def hdr_update(self):
+        self.video_options.advanced.hdr_settings()
+        self.encoder_settings_update()
 
     def page_update(self, build_thumbnail=True):
         if not self.initialized or self.loading_video or not self.app.fastflix.current_video:
@@ -1516,7 +1539,7 @@ class Main(QtWidgets.QWidget):
         video = self.app.fastflix.current_video
 
         self.app.fastflix.queue.append(copy.deepcopy(video))
-        self.video_options.update_queue()
+        self.video_options.update_queue(currently_encoding=self.converting)
         self.video_options.show_queue()
 
         if self.converting:
@@ -1546,26 +1569,30 @@ class Main(QtWidgets.QWidget):
         self.set_convert_button()
 
         try:
-            video_uuid, command_uuid = data.split(":")
+            video_uuid, command_uuid, *_ = data.split("|")
             cancelled_video = self.find_video(video_uuid)
         except Exception:
             return
 
-        sm = QtWidgets.QMessageBox()
-        sm.setWindowTitle(t("Cancelled"))
-        sm.setText(
-            f"{t('Conversion cancelled, delete incomplete file')}\n" f"{cancelled_video.video_settings.output_path}?"
-        )
-        sm.addButton(t("Delete"), QtWidgets.QMessageBox.YesRole)
-        sm.addButton(t("Keep"), QtWidgets.QMessageBox.NoRole)
-        sm.exec_()
-        if sm.clickedButton().text() == t("Delete"):
-            try:
-                video_uuid, command_uuid = data.split(":")
-                cancelled_video = self.find_video(video_uuid)
-                cancelled_video.video_settings.output_path.unlink(missing_ok=True)
-            except OSError:
-                pass
+        if self.video_options.queue.paused:
+            self.video_options.queue.pause_resume_queue()
+
+        if cancelled_video.video_settings.output_path.exists():
+            sm = QtWidgets.QMessageBox()
+            sm.setWindowTitle(t("Cancelled"))
+            sm.setText(
+                f"{t('Conversion cancelled, delete incomplete file')}\n"
+                f"{cancelled_video.video_settings.output_path}?"
+            )
+            sm.addButton(t("Delete"), QtWidgets.QMessageBox.YesRole)
+            sm.addButton(t("Keep"), QtWidgets.QMessageBox.NoRole)
+            sm.exec_()
+            if sm.clickedButton().text() == t("Delete"):
+                try:
+                    cancelled_video = self.find_video(video_uuid)
+                    cancelled_video.video_settings.output_path.unlink(missing_ok=True)
+                except OSError:
+                    pass
 
     @reusables.log_exception("fastflix", show_traceback=False)
     def dropEvent(self, event):
@@ -1578,13 +1605,20 @@ class Main(QtWidgets.QWidget):
             self.input_video = Path(event.mimeData().urls()[0].toLocalFile())
         except (ValueError, IndexError):
             return event.ignore()
-        else:
-            self.video_path_widget.setText(str(self.input_video))
-            self.output_video_path_widget.setText(self.generate_output_filename)
-            self.output_video_path_widget.setDisabled(False)
-            self.output_path_button.setDisabled(False)
+
+        self.video_path_widget.setText(str(self.input_video))
+        self.output_video_path_widget.setText(self.generate_output_filename)
+        self.output_video_path_widget.setDisabled(False)
+        self.output_path_button.setDisabled(False)
+        try:
             self.update_video_info()
-            self.page_update()
+        except Exception:
+            logger.exception(f"Could not load video {self.input_video}")
+            self.video_path_widget.setText("")
+            self.output_video_path_widget.setText("")
+            self.output_video_path_widget.setDisabled(True)
+            self.output_path_button.setDisabled(True)
+        self.page_update()
 
     def dragEnterEvent(self, event):
         event.accept() if event.mimeData().hasUrls else event.ignore()
@@ -1595,7 +1629,7 @@ class Main(QtWidgets.QWidget):
     def status_update(self, status):
         logger.debug(f"Updating status from command worker: {status}")
         try:
-            command, video_uuid, command_uuid = status.split(":")
+            command, video_uuid, command_uuid, *_ = status.split("|")
         except ValueError:
             logger.exception(f"Could not process status update from the command worker: {status}")
             return
@@ -1616,24 +1650,24 @@ class Main(QtWidgets.QWidget):
                     if mkv_prop_edit := shutil.which("mkvpropedit"):
                         worker = SubtitleFix(self, mkv_prop_edit, video.video_settings.output_path)
                         worker.start()
-                self.video_options.update_queue()
+                self.video_options.update_queue(currently_encoding=self.converting)
             else:
                 logger.error(f"This should not happen? {status} - {video}")
 
         elif command == "running":
             video.status.current_command = command_index
             video.status.running = True
-            self.video_options.update_queue(currently_encoding=True)
+            self.video_options.update_queue(currently_encoding=self.converting)
 
         elif command == "error":
             video.status.error = True
             video.status.running = False
-            self.video_options.update_queue()
+            self.video_options.update_queue(currently_encoding=self.converting)
 
         elif command == "cancelled":
             video.status.cancelled = True
             video.status.running = False
-            self.video_options.update_queue()
+            self.video_options.update_queue(currently_encoding=self.converting)
 
         elif command in ("paused encode", "resumed encode"):
             pass
@@ -1670,11 +1704,11 @@ class Notifier(QtCore.QThread):
             if status[0] == "complete":
                 self.main.completed.emit(0)
             elif status[0] == "error":
-                self.main.status_update_signal.emit(":".join(status))
+                self.main.status_update_signal.emit("|".join(status))
                 self.main.completed.emit(1)
             elif status[0] == "cancelled":
-                self.main.cancelled.emit(":".join(status[1:]))
-                self.main.status_update_signal.emit(":".join(status))
+                self.main.cancelled.emit("|".join(status[1:]))
+                self.main.status_update_signal.emit("|".join(status))
             elif status[0] == "exit":
                 try:
                     self.terminate()
@@ -1682,4 +1716,4 @@ class Notifier(QtCore.QThread):
                     self.main.close_event.emit()
                 return
             else:
-                self.main.status_update_signal.emit(":".join(status))
+                self.main.status_update_signal.emit("|".join(status))
